Fix a few more tests.

+++ NeuralObscura/Shaders.metal
@@ -28,29 +28,29 @@ kernel void rectifier_linear(texture2d_array<float, access::read> inTexture [[te
  *
  * Formula: output = (gamma * ((input - mean) / stddev)) + beta
  */
-kernel void batch_normalization(texture2d_array<float, access::read> inTexture [[texture(0)]],
-                                texture2d_array<float, access::write> outTexture [[texture(1)]],
+kernel void batch_normalization(texture2d_array<half, access::read> inTexture [[texture(0)]],
+                                texture2d_array<half, access::write> outTexture [[texture(1)]],
                                 const device float* gamma [[ buffer(2) ]],
                                 const device float* beta [[ buffer(3) ]],
                                 const device float* mean [[ buffer(4) ]],
                                 const device float* stddev [[ buffer(5) ]],
                                 uint3 gid [[thread_position_in_grid]]) {
     uint buffer_idx = gid.z * 4;
-    float4 input = inTexture.read(gid.xy, gid.z);
+    half4 input = inTexture.read(gid.xy, gid.z);
     float no_zero_divide = 0.0000001;
-    float4 output = float4(input.r - mean[buffer_idx],
+    half4 output = half4(input.r - mean[buffer_idx],
                            input.g - mean[buffer_idx+1],
                            input.b - mean[buffer_idx+2],
                            input.a - mean[buffer_idx+3]);
-    output = float4(output.r / (stddev[buffer_idx] + no_zero_divide),
+    output = half4(output.r / (stddev[buffer_idx] + no_zero_divide),
                     output.g / (stddev[buffer_idx+1] + no_zero_divide),
                     output.b / (stddev[buffer_idx+2] + no_zero_divide),
                     output.a / (stddev[buffer_idx+3] + no_zero_divide));
-    output = float4(output.r * gamma[buffer_idx],
+    output = half4(output.r * gamma[buffer_idx],
                     output.g * gamma[buffer_idx+1],
                     output.b * gamma[buffer_idx+2],
                     output.a * gamma[buffer_idx+3]);
-    output = float4(output.r + beta[buffer_idx],
+    output = half4(output.r + beta[buffer_idx],
                     output.g + beta[buffer_idx+1],
                     output.b + beta[buffer_idx+2],
                     output.a + beta[buffer_idx+3]);
